// src/app/auxiliar/models/tabla-auxiliar-detalle.ts
import { TablaAuxiliarDetalleId } from "./tabla-auxiliar-detalle-id";

export class TablaAuxiliarDetalle {
    tablaAuxiliarDetalleId: TablaAuxiliarDetalleId;
    nombre: string;
    abreviatura: string;
    valor: string;
    valor2: string;
    observacion: string;
    indHabilitado: string;
    idUsuarioCrea: number;
    fechaCrea: Date;
}